dashboard_api/requirements.txt
fastapi
uvicorn[standard]
python-socketio
asyncpg
redis
stripe
python-dotenv
